requirements.txt
# ─── UI ──────────────────────────────────────────────────────────────────────
streamlit>=1.36.0

# ─── Resume Parsing ───────────────────────────────────────────────────────────
pdfplumber>=0.11.2
python-docx>=1.1.2

# ─── NLP / NER ────────────────────────────────────────────────────────────────
spacy>=3.7.5
# After install, run: python -m spacy download en_core_web_sm

# ─── Data ────────────────────────────────────────────────────────────────────
pandas>=2.2.2

# ─── PDF Report Export ────────────────────────────────────────────────────────
fpdf2>=2.7.9

# ─── HTTP ─────────────────────────────────────────────────────────────────────
requests>=2.31.0

# ─── Optional: Firebase ───────────────────────────────────────────────────────
# firebase-admin>=6.4.0
